rag/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sentence-transformers==2.7.0
PyPDF2==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
transformers==4.42.3
torch==2.3.1